fix: V-002 security vulnerability
Automated security fix generated by OrbisAI Security

+++ shadowsocksr-libev/src/libev/ev.c
@@ -1946,7 +1946,7 @@ static void * noinline ecb_cold
 array_realloc (int elem, void *base, int *cur, int cnt)
 {
   *cur = array_nextsize (elem, *cur, cnt);
-  return ev_realloc (base, elem * *cur);
+  return ev_realloc (base, (size_t)elem * (size_t)*cur);
 }
 
 #define array_init_zero(base,count)	\
